Refactor to functions.
