+++ src/useSubscription/models/subscription.ts
@@ -3,7 +3,7 @@ import { ref, Ref, watch } from 'vue'
 import Channel from '@/useSubscription/models/channel'
 import { Action, ActionResponse } from '@/useSubscription/types/action'
 import { SubscriptionOptions } from '@/useSubscription/types/subscription'
-import { useSubscriptionDevtoolsInspector } from '../devtools'
+import * as useSubscriptionDevtools from '@/useSubscription/useSubscriptionDevtools'
 
 class SubscriptionIdManager {
   private static id: number = 0
@@ -42,7 +42,7 @@ export default class Subscription<T extends Action> {
 
   public unsubscribe(): void {
     this.channel.unsubscribe(this.id)
-    useSubscriptionDevtoolsInspector.removeChannelSubscription(this.channel, this.id)
+    useSubscriptionDevtools.removeChannelSubscription(this.channel, this.id)
   }
 
   public isSubscribed(): boolean {
